api/src/requirements.txt
numpy==1.19.1
google-api-python-client==1.10.1
sklearn==0.0
Flask==1.1.2
Keras-Preprocessing==1.1.2
google-api-core==1.22.1
gunicorn==20.0.4
six==1.15.0
python-decouple==3.3